requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-experimental==0.3.0
pandas==2.2.0
openpyxl==3.1.4
langchain-openai==0.3.0
faiss-cpu==1.9.0
PyMuPDF==1.25.0
python-dotenv==1.0.1
streamlit==1.41.0
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.19
langchain-postgres==0.0.12
psycopg==3.2.3
psycopg-binary==3.2.3